+++ common/reactive/src/main/java/is/codion/common/reactive/value/AbstractValueCollection.java
@@ -203,7 +203,7 @@ abstract class AbstractValueCollection<T, C extends Collection<T>> extends BaseV
 	}
 
 	@Override
-	public ObservableValueCollection<T, C> observable() {
+	public synchronized ObservableValueCollection<T, C> observable() {
 		return (ObservableValueCollection<T, C>) super.observable();
 	}